Rename typeql -> query behaviour packages
