+++ test/behaviour/connection/session/BUILD
@@ -56,7 +56,7 @@ java_test(
         "//test/behaviour/config:parameters",
         "//test/behaviour/connection/database:steps",
         "//test/behaviour/connection/transaction:steps",
-        "//test/behaviour/typeql:steps",
+        "//test/behaviour/query:steps",
     ],
     data = [
         "@vaticle_typedb_behaviour//connection:session.feature",
